requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
boto3>=1.34.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
beautifulsoup4>=4.12.2
requests>=2.31.0
html5lib>=1.1
python-dotenv>=1.0.0
jinja2>=3.1.2
pillow>=10.0.0
torch>=2.0.0